Coming back from buster backports due to end of support
We previously used backports packages to fix in emergency some apache2 security issues that were not fixed at that time on the standard packages. The use of backports is not required anymore and backports is no longer available on the mirrors. 
Debian Backports does not support LTS [1], therefore buster-backports is unsupported since August 1st 2022 but was still available on the mirrors, that changed recently with the archival of buster-backports.
 see https://backports.debian.org/news/Removal_of_buster-backports_from_the_debian_archive/